feat: add new scenes config page support
