+++ shell/app/common/components/load-more-selector.scss
@@ -144,7 +144,7 @@
         margin-top: 16px;
       }
 
-      .list-item {
+      .load-more-list-item {
         display: flex;
         align-items: center;
         margin-left: 0;
